+++ features/admin.flow-builder-core.v1/models/steps.ts
@@ -65,5 +65,6 @@ export enum StaticStepTypes {
 }
 
 export enum RedirectionTypes {
-    GoogleFederation = "GoogleExecutor"
+    GoogleFederation = "GoogleExecutor",
+    PasskeyEnrollment = "FIDO2Executor"
 }